requirements.txt
ctransformers
sentence-transformers
pinecone-client
langchain_pinecone
langchain_community
flask
pypdf
python-dotenv

-e .